Moving GuiEvent to event.py
